pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "3D-pose-estimation"
version = "0.1.0"
description = "Real-time 3D human pose estimation from RGB-D using MediaPipe + OpenCV + Open3D"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
  "opencv-python>=4.9.0",
  "mediapipe>=0.10.14",
  "numpy>=1.26",
  "open3d>=0.18.0",
  "matplotlib>=3.8",
  "pyyaml>=6.0.1",
  "tqdm>=4.66",
  "scipy>=1.11",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "black>=24.3.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "types-PyYAML",
  "types-tqdm",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]
src = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["src/pose3d"]